Added safeguards
Improve the safety of the user input by adding options where if the
input is incorrect it requests the option again. It doesn’t quit just
because of an incorrect input.
Also we added different options but still missing the sub menus and get
the actions to execute in ssh per selected action